Move PointerBlock to History (#RCHAIN-3415)
