+++ rspace/src/main/scala/coop/rchain/rspace/nextgenrspace/history/History.scala
@@ -53,6 +53,41 @@ object Trie {
       }(codecLeaf.as[Leaf])
 }
 
+final case class PointerBlock private (toVector: Vector[Trie]) {
+  def updated(tuples: List[(Int, Trie)]): PointerBlock =
+    new PointerBlock(tuples.foldLeft(toVector) { (vec, curr) =>
+      vec.updated(curr._1, curr._2)
+    })
+
+  def countNonEmpty: Int = toVector.count(_ != EmptyTrie)
+}
+
+object PointerBlock {
+  val length = 256
+
+  def create(): PointerBlock = new PointerBlock(Vector.fill(length)(EmptyTrie))
+
+  def create(first: (Int, Trie)): PointerBlock =
+    PointerBlock.create().updated(List(first))
+
+  def create(first: (Int, Trie), second: (Int, Trie)): PointerBlock =
+    PointerBlock.create().updated(List(first, second))
+
+  implicit val codecPointerBlock: Codec[PointerBlock] =
+    vectorOfN(
+      provide(length),
+      Trie.codecTrie
+    ).as[PointerBlock]
+
+  def hash(pb: PointerBlock): Blake2b256Hash =
+    codecPointerBlock
+      .encode(pb)
+      .map((vector: BitVector) => Blake2b256Hash.create(vector.toByteArray))
+      .get
+
+  def unapply(arg: PointerBlock): Option[Vector[Trie]] = Option(arg.toVector)
+}
+
 final case class TriePath(nodes: Vector[Trie], conflicting: Option[Trie], edges: KeyPath) {
   def append(affix: KeyPath, t: Trie): TriePath =
     this.copy(nodes = this.nodes :+ t, edges = this.edges ++ affix)
